+++ restli-server/src/main/java/com/linkedin/restli/internal/server/ResourceContextImpl.java
@@ -611,16 +611,6 @@ public class ResourceContextImpl implements ServerResourceContext
     return _responseStreamingAttachments;
   }
 
-  /**
-   * @deprecated Use {@link #isReturnEntityRequested()} instead.
-   */
-  @Deprecated
-  @Override
-  public boolean shouldReturnEntity()
-  {
-    return isReturnEntityRequested();
-  }
-
   @Override
   public boolean isReturnEntityRequested()
   {